Improve metadata documentation
Add metadata field documentation to Propose Agreement, Propose
Unassigned Agreement and Create Open Agreement.

Mention that the metadata is stored against transactions and included
in the transaction webhook payloads for Payouts and Payment Requests.

Remove blanket statement alluding to the above for all Metadata as it
only applies to Payouts and Payment Requests.